+++ changelog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2019-09-29
+### Added
+- Kill switch in `exit()`.
+- Rough version of `bones.wav`.
+- Audio trigger.
+
+### Changed
+- Made `red()` more efficient.
+
 ## [1.0.1] - 2019-09-28
 ### Added
 - Changelog template.